requirements.txt
langchain
openai
pyyaml